+++ Changelog.md
@@ -46,6 +46,11 @@ Deprecations:
   CLI). Use location filtering instead. (Myron Marston)
 * Deprecate `--default_path` as an alternative to `--default-path`. (Jon Rowe)
 
+Bug Fixes:
+
+* Issue a warning when you set `config.deprecation_stream` too late for
+  it to take effect because the reporter has already been setup. (Myron Marston)
+
 ### 2.99.0.beta2 / 2014-02-17
 [full changelog](http://github.com/rspec/rspec-core/compare/v2.99.0.beta1...v2.99.0.beta2)